Move contribution lookups to their own module
