+++ src/github.rs
@@ -3,18 +3,11 @@
 
 use anyhow::Context;
 use chrono::{DateTime, Utc};
-use indicatif::ProgressBar;
 use isahc::prelude::*;
 use serde::{Deserialize, Serialize};
 
-/// The maximum number of results to fetch in a page.
-const CONTRIBUTION_PAGE_SIZE: u32 = 5;
-
-/// The maximum number of pages to pull when querying for user contributions.
-const CONTRIBUTION_MAX_PAGES: u32 = 10 * (100 / CONTRIBUTION_PAGE_SIZE);
-
 /// The never-changing URL to POST to for any V4 request.
-const V4_URL: &str = "https://api.github.com/graphql";
+pub const V4_URL: &str = "https://api.github.com/graphql";
 
 const LIMIT_QUERY: &str = "{ \
   \"query\": \"{ \
@@ -41,53 +34,10 @@ struct RateLimitQuery {
 }
 
 #[derive(Debug, Deserialize)]
-struct SearchQuery {
-    search: Paged<UserContributions>,
-}
-
-#[derive(Debug, Deserialize)]
 #[serde(rename_all = "camelCase")]
-pub struct UserContributions {
-    pub login: String,
-    pub name: Option<String>,
-    pub followers: Followers,
-    pub contributions_collection: Contributions,
-}
-
-impl UserContributions {
-    pub fn contributions(&self) -> u32 {
-        let total = self
-            .contributions_collection
-            .contribution_calendar
-            .total_contributions;
-        total - self.contributions_collection.restricted_contributions_count
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct Followers {
-    pub total_count: u32,
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct Contributions {
-    pub contribution_calendar: Calendar,
-    pub restricted_contributions_count: u32,
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct Calendar {
-    pub total_contributions: u32,
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct PageInfo {
-    has_next_page: bool,
-    end_cursor: Option<String>,
+pub struct PageInfo {
+    pub has_next_page: bool,
+    pub end_cursor: Option<String>,
 }
 
 #[derive(Debug, Deserialize)]
@@ -102,9 +52,9 @@ pub struct Node<A> {
 
 #[derive(Debug, Deserialize)]
 #[serde(rename_all = "camelCase")]
-struct Paged<A> {
-    page_info: PageInfo,
-    edges: Vec<Node<A>>,
+pub struct Paged<A> {
+    pub page_info: PageInfo,
+    pub edges: Vec<Node<A>>,
 }
 
 /// A single structure that represents the results from either an `issues` call
@@ -141,8 +91,8 @@ pub struct CommitCount {
 
 /// The top-level results of a GraphQL query.
 #[derive(Debug, Deserialize)]
-struct Query<T> {
-    data: T,
+pub struct Query<T> {
+    pub data: T,
 }
 
 #[derive(Deserialize)]
@@ -198,42 +148,6 @@ impl Mode {
             _ => "",
         }
     }
-}
-
-fn users_query(location: &str, page: Option<&str>) -> String {
-    format!(
-        "{{ \
-         \"query\": \"{{ \
-         search(type: USER, query: \\\"type:user location:{} sort:followers-desc\\\", first: {}{}) {{ \
-           pageInfo {{ \
-             hasNextPage \
-             endCursor \
-           }} \
-           edges {{ \
-             node {{ \
-               ... on User {{ \
-                 login \
-                 name \
-                 followers {{ \
-                   totalCount \
-                 }} \
-                 contributionsCollection {{ \
-                   contributionCalendar {{ \
-                     totalContributions \
-                   }} \
-                   restrictedContributionsCount \
-                 }} \
-               }} \
-             }} \
-           }} \
-         }} \
-       }}\" \
-    }}",
-        location,
-        CONTRIBUTION_PAGE_SIZE,
-        page.map(|p| format!(", after: \\\"{}\\\"", p))
-            .unwrap_or_else(|| "".to_string()),
-    )
 }
 
 fn issue_query(mode: &Mode, owner: &str, repo: &str, page: Option<&str>) -> String {
@@ -344,52 +258,6 @@ pub fn rate_limit(client: &HttpClient) -> anyhow::Result<RateLimit> {
         .with_context(|| format!("The response couldn't be decoded into JSON:\n{}", text))?;
 
     Ok(limit_query.data.rate_limit)
-}
-
-/// Produce a list of Github Users, ordered by their contribution counts.
-pub fn user_contributions(
-    client: &HttpClient,
-    location: &str,
-) -> anyhow::Result<Vec<UserContributions>> {
-    let bar = ProgressBar::new(CONTRIBUTION_MAX_PAGES as u64);
-    bar.set_message("Fetching User contributions...");
-    let result = user_contributions_work(client, &bar, location, None, 1);
-    bar.finish_and_clear();
-    result
-}
-
-fn user_contributions_work(
-    client: &HttpClient,
-    bar: &ProgressBar,
-    location: &str,
-    page: Option<&str>,
-    page_num: u32,
-) -> anyhow::Result<Vec<UserContributions>> {
-    bar.inc(1);
-
-    let body = users_query(location, page);
-
-    let mut resp = client
-        .post(V4_URL, body)
-        .context("There was a problem calling the Github GraphQL API.")?;
-
-    let text = resp.text()?;
-
-    let result: Query<SearchQuery> = serde_json::from_str(&text)
-        .with_context(|| format!("The response couldn't be decoded into JSON:\n{}", text))?;
-
-    let page = result.data.search;
-    let info = page.page_info;
-    let mut users: Vec<UserContributions> = page.edges.into_iter().map(|n| n.node).collect();
-
-    match info.end_cursor {
-        Some(c) if info.has_next_page && page_num < CONTRIBUTION_MAX_PAGES => {
-            let mut next = user_contributions_work(client, bar, location, Some(&c), page_num + 1)?;
-            users.append(&mut next);
-            Ok(users)
-        }
-        _ => Ok(users),
-    }
 }
 
 #[derive(Debug, Deserialize)]
