+++ src/lib.rs
@@ -1,10 +1,11 @@
 //! A library for measuring Github repository contributions.
 
+mod contributions;
 mod github;
 
 // Re-export.
+pub use contributions::user_contributions;
 pub use github::rate_limit;
-pub use github::user_contributions;
 
 use anyhow::Context;
 use chrono::{DateTime, Utc};
